requirements.txt
Django>=5.0,<6.0

djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.3,<6.0

gunicorn>=21.0,<22.0
whitenoise>=6.6,<7.0

django-cors-headers>=4.3,<5.0
python-dotenv>=1.0,<2.0

django-filter>=24.2,<25.0

requests>=2.32,<3.0